feat: create seedlot number api (#440)
* feat: create next seedlot number

Create endpoint, dto and service

About issue #237

* feat: seedlot response and saving - partial

* feat: next seedlot number constants and query

* feat: add logs and finish service create seedlot

* chore: fix code style formatting

* tests: add user service tests

* tests: add seedlot endpoint tests - partial

* tests: add seedlot endpoint test

* tests: add seedlot service test

* tests: add seedlot repository test - partial

* tests: add seedlot repository test

* feat: add genetic class code to the seedlot dto

Issue #237

Add geneticClassCode to the SeedlotCreateDto class, which should be
filled with 'A' or 'B', a single character representing the seedlot
class code. For now this field is optional, since the Seedlot table
allows this columns to be null, which means we can populate later.

* feat: make genetic class code field required

Issue #237

All requests for registering a seedlot should now include a genetic
class code representing the class. E.g.: A or B.

* chore: fix possible null value warning

No issues

This change is not related to other changes on issue #237 but to a small
improvement on code warning.

* tests: add more test cases for missing entities

Issue #237

Note that all method names were rewritten, given a common convention for
test method names. And since these method names can get big, we can use
the DisplayName to provide a better description on console, when
failing. These names convention were brought by Om on a small discussion
on FSADT2 Discord channel, thanks Om :)

---------