feat: add effective population size calculation (#468)
* feat: add method for effective population size ne

Issue #238

Creates a private method in the Genetic Worth Service that will do all
the calculations related to Effective Population Size (Ne). Note that
for now, this method is considering the UNTESTED approach, which is much
more simpler than the TESTED one. The TESTED one should be tackled
later.

* feat: add example value for Ne summary dto

Issue #238

Add a sample value that would be returned for an effective population
size calculation.

---------